# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: TutorLink API
  version: 0.1.0
  description: Foundation contract for Auth, File, and Marketplace Listing flows.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
servers:
  - url: http://127.0.0.1:8000
    description: Local Docker development
...
paths:
...
  /api/v1/auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [Auth]
      summary: Request a password reset link
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset request accepted.
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/v1/auth/reset-password:
    post:
      operationId: resetPassword
      tags: [Auth]
      summary: Reset a password with a valid token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password reset completed.
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/auth/email/verify/{id}/{hash}:
    get:
      operationId: verifyEmail
      tags: [Auth]
      summary: Verify the authenticated user's email address
      security: [{ sanctumCookie: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, minimum: 1 }
        - name: hash
          in: path
          required: true
          schema: { type: string }
        - name: expires
          in: query
          required: true
          schema: { type: integer }
        - name: signature
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Email verified.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/auth/email/verification-notification:
    post:
      operationId: resendVerification
      tags: [Auth]
      summary: Resend the email verification notification
      security: [{ sanctumCookie: [] }]
      responses:
        "200":
          description: Verification notification dispatched.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /api/v1/files/{file}:
    parameters:
      - $ref: "#/components/parameters/FileId"
    get:
      operationId: getFile
      tags: [Files]
      summary: Return file metadata when the viewer is authorized
      security: [{ sanctumCookie: [] }]
      responses:
        "200":
          description: File metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileResponse"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteFile
      tags: [Files]
      summary: Remove an unlinked file owned by the current user
      security: [{ sanctumCookie: [] }]
      responses:
        "204":
          description: File removed.
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: File is linked to a business record.
...
components:
  securitySchemes:
    sanctumCookie:
      type: apiKey
      in: cookie
      name: tutorlink-session
  parameters:
    FileId:
      name: file
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
    Unauthenticated:
      description: Authentication is required.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: The authenticated user is not allowed to perform this action.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ValidationError" }
    TooManyRequests:
      description: Too many requests.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    User:
      type: object
      required: [id, role, status, full_name, email]
      properties:
        id: { type: integer, format: int64, minimum: 1 }
        role: { type: string, enum: [customer, tutor, admin] }
        status: { type: string, enum: [pending, active, suspended, cancelled] }
        full_name: { type: string, maxLength: 255 }
        email: { type: string, format: email }
        date_of_birth: { type: string, format: date, nullable: true }
        email_verified_at: { type: string, format: date-time, nullable: true }
    UserResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/User" }
...
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    ResetPasswordRequest:
      type: object
      required: [token, email, password, password_confirmation]
      properties:
        token: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
        password_confirmation: { type: string, format: password }
...
    File:
      type: object
      required:
        [
          id,
          owner_id,
          original_name,
          mime_type,
          size_bytes,
          scan_status,
          purpose,
          complete,
          created_at,
        ]
      properties:
        id: { type: integer, format: int64, minimum: 1 }
        owner_id: { type: integer, format: int64, minimum: 1 }
        original_name: { type: string, nullable: true }
        mime_type: { type: string }
        size_bytes: { type: integer, format: int64, minimum: 0 }
        purpose:
          {
            type: string,
            enum: [avatar, verification_document, task_attachment, profile_document],
          }
        scan_status: { type: string, enum: [pending, clean, rejected] }
        complete: { type: boolean }
        created_at: { type: string, format: date-time }
    FileResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/File" }
...
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        request_id: { type: string, format: uuid, nullable: true }
    ValidationError:
      allOf:
        - $ref: "#/components/schemas/Error"
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: array
                items: { type: string }